+++ Application/Shared/DTOs/RegisterModality/RegisterModalityDto.cs
@@ -2,11 +2,11 @@ using System;
 
 namespace Application.Shared.DTOs.RegisterModality
 {
-    public class RegisterModalityDto : BaseDto<long>
+    public class RegisterModalityDto : BaseDto<int>
     {
-        public long IdModality { get; set; }
-        public long IdRegisterModalityState { get; set; }
-        public long IdAcademicPeriod { get; set; }
+        public int IdModality { get; set; }
+        public int IdRegisterModalityState { get; set; }
+        public int IdAcademicPeriod { get; set; }
         public DateTime? ApprovalDate { get; set; }
         public string? Observations { get; set; }
         public new int? IdUserCreatedAt { get; set; }
